+++ components/autofill/core/browser/ui/country_combobox_model.h
@@ -1,4 +1,4 @@
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2012 The Chromium Authors
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
@@ -9,7 +9,7 @@
 #include <string>
 #include <vector>
 
-#include "base/callback.h"
+#include "base/functional/callback.h"
 #include "ui/base/models/combobox_model.h"
 
 namespace autofill {
@@ -31,6 +31,7 @@ class CountryComboboxModel : public ui::ComboboxModel {
 
   // |filter| is passed each known country's country code. If |filter| returns
   // true, an item for that country is added to the model (else it's omitted).
+  // Empty callback can be used to retain all countries.
   // |manager| determines the default choice.
   void SetCountries(
       const PersonalDataManager& manager,
@@ -38,9 +39,9 @@ class CountryComboboxModel : public ui::ComboboxModel {
       const std::string& app_locale);
 
   // ui::ComboboxModel implementation:
-  int GetItemCount() const override;
-  std::u16string GetItemAt(int index) const override;
-  bool IsItemSeparatorAt(int index) const override;
+  size_t GetItemCount() const override;
+  std::u16string GetItemAt(size_t index) const override;
+  bool IsItemSeparatorAt(size_t index) const override;
 
   // The list of countries always has the default country at the top as well as
   // within the sorted vector.
